Move the chess tests into tests/chess/
Everything in tests/ except conformance.py is driven by chess positions, so at
three games it would have been fifteen flat files with nothing to say who owns
which. tests/ now mirrors games/.

The search and simulation tests move too, even though they exercise ai/ rather
than chess: chess is where their fixtures come from, and splitting them out
would have put chess positions in two places. When a second game arrives it
brings tests/<name>/ and one import block in test_all.py, and touches nothing
here.
