+++ README.md
@@ -74,7 +74,7 @@ Search time grows steeply with depth — see the benchmarks below.
 | `uci/engine.py` | Mildred's own UCI server (`main.py` runs this) |
 | `uci/protocol.py`, `uci/stockfish.py` | An async UCI *client*, for driving other engines |
 | `tests/conformance.py` | The tests every game must pass |
-| `tests/` | Unit tests, full-game replay/undo tests, perft |
+| `tests/chess/` | Rules, replay/undo, perft, and search scores, all driven by chess positions |
 
 Positions are held as twelve 64-bit integers (one per piece type per colour) plus occupancy
 masks. Sliding attacks are generated by masking precomputed rays against blockers; knight,
@@ -129,8 +129,8 @@ Verified against the [reference results](https://www.chessprogramming.org/Perft_
 | Position 4 | 3 | 9,467 | ✅ |
 | Position 5 | 3 | 62,379 | ✅ |
 
-All five run as part of the test suite (`tests/test_permutations.py`), so move generation is
-pinned to the reference counts rather than merely known to have matched them once.
+All five run as part of the test suite (`tests/chess/test_permutations.py`), so move generation
+is pinned to the reference counts rather than merely known to have matched them once.
 
 ## Playing against other engines
 
@@ -156,15 +156,15 @@ extras either.
 Run the suites individually while iterating:
 
 ```bash
-python3 -m unittest tests.test_moves -v               # rules and legality
-python3 -m unittest tests.test_undo -v                # full-game replay, forwards and back
-python3 -m unittest tests.test_permutations -v        # perft
-python3 -m unittest tests.test_conformance_chess -v   # chess against the shared game contract
-python3 -m unittest tests.test_search_equivalence -v  # search scores, position by position
+python3 -m unittest tests.chess.test_moves -v               # rules and legality
+python3 -m unittest tests.chess.test_undo -v                # full-game replay, forwards and back
+python3 -m unittest tests.chess.test_permutations -v        # perft
+python3 -m unittest tests.chess.test_conformance -v         # chess against the shared game contract
+python3 -m unittest tests.chess.test_search_equivalence -v  # search scores, position by position
 ```
 
-`tests/test_search_equivalence.py` is the net for changing the search, which has no oracle but
-itself. It scores every root move over a reproducible corpus rather than only the move that
-wins — two searches can agree on the best move and disagree about everything else. Widen
-`CORPUS_SIZE` and `DEPTHS` there before touching the search, and expect minutes: cost grows as
-branching^depth.
+`tests/chess/test_search_equivalence.py` is the net for changing the search, which has no
+oracle but itself. It scores every root move over a reproducible corpus rather than only the
+move that wins — two searches can agree on the best move and disagree about everything else.
+Widen `CORPUS_SIZE` and `DEPTHS` there before touching the search, and expect minutes: cost
+grows as branching^depth.
