+++ tests/chess/corpus.py
