+++ tests/chess/test_conformance.py
