+++ tests/chess/test_permutations.py
