+++ tests/chess/test_simulate.py
