+++ tests/test_all.py
@@ -1,13 +1,23 @@
+"""
+The whole suite.
+
+Tests live beside the game whose positions they use: `tests/chess/` for everything driven by a
+chess board, including the tests of `ai/` that use chess as their vehicle. Only what is shared
+between games stays here - `tests/conformance.py`, the contract every game must satisfy.
+
+Adding a game means adding a `tests/<name>/` package and one import block below.
+"""
+
 import unittest
 
-from tests.test_board import TestBitboard
-from tests.test_moves import TestMoves
-from tests.test_undo import TestUndo
-from tests.test_permutations import TestPermutations
-from tests.test_search import TestSearch, TestTerminalValue, TestUciSearchOutput
-from tests.test_search_equivalence import TestHarness, TestDecisivePositions
-from tests.test_simulate import TestResult, TestSimulateGame
-from tests.test_conformance_chess import TestChessConformance
+from tests.chess.test_board import TestBitboard
+from tests.chess.test_moves import TestMoves
+from tests.chess.test_undo import TestUndo
+from tests.chess.test_permutations import TestPermutations
+from tests.chess.test_search import TestSearch, TestTerminalValue, TestUciSearchOutput
+from tests.chess.test_search_equivalence import TestHarness, TestDecisivePositions
+from tests.chess.test_simulate import TestResult, TestSimulateGame
+from tests.chess.test_conformance import TestChessConformance
 
 
 def main():
